Use Album Artist field, split artist field, refactor final structure.
If available, use Album Artist field instead of artist.
Split Artist field when featured, and or comma patern is present.
If album artist differs artist, add the song to all artists in the array.
This is to separate "B.B.King and Eric Clapton" from "Derec and the Dominoes" and the likes.
Better structure for the database, get rid of "All songs" and use only "All - full name".